fix(config): tighten active_leg_tier fields to uint16
Mirrors the Rust u16 pair on ActiveLegTier (eligibility.rs). Rejects out-of-range min_active_legs/max_commission_depth at Go YAML unmarshal instead of truncating at the Rust boundary. Adds boundary tests (65535 ok; 65536/70000/-1 rejected).

+++ internal/config/types.go
@@ -197,8 +197,8 @@ type CommissionEligibility struct {
 
 // ActiveLegTier maps an active leg count to a commission depth.
 type ActiveLegTier struct {
-	MinActiveLegs      int `yaml:"min_active_legs" json:"min_active_legs"`
-	MaxCommissionDepth int `yaml:"max_commission_depth" json:"max_commission_depth"`
+	MinActiveLegs      uint16 `yaml:"min_active_legs" json:"min_active_legs"`
+	MaxCommissionDepth uint16 `yaml:"max_commission_depth" json:"max_commission_depth"`
 }
 
 // --- Commission interface ---
